Split out few things from the Audio's manager into smaller submodules
