+++ redbot/cogs/audio/core/abc.py
@@ -35,7 +35,7 @@ if TYPE_CHECKING:
     from ..apis.playlist_wrapper import PlaylistWrapper
     from ..audio_dataclasses import LocalPath, Query
     from ..equalizer import Equalizer
-    from ..manager import ServerManager
+    from ..manager.server_manager import ServerManager
 
 
 class MixinMeta(ABC):
